Add a regression test script
This is currently very simple: it just checks out and builds the
current repository with the default configuration and then runs all
the ISA tests on Spike.  The general idea here is to put most of the
regression logic in this script, both so users can run it and so the
various regression systems don't end up with all this logic baked in.