Broken in-document links, mistake, style fix
There are many broken links in the third part of the manual,
which are fixed now.
Right before the "Help!" section, there was a command:
	spike +disk=root.bin vmlinux
which should be
	spike +disk=root.bin bbl vmlinux
and now it is fixed as well.

Finally about style, I think that the references provided originally
in subsection 10 of Linux manual are actually of a whole-project
scope, and thus can be move up a level.