Add missing dependency: zlib1g-dev
zlib1g-dev is listed in the quick start section, but not in this location.  If you skip the quick start and only install the packages listed in the "Obtaining and Compiling the Sources" section, ./build.sh will fail due to missing zlib (on systems that don't have this library already -- my fresh Ubuntu 16.04 vmware install did not).